Move node/* python files into a calico package.
Code rearrangement to move all of our Python code into
a single package, for ease of commonization.
