+++ calico/adapter/datastore.py
