+++ calico/adapter/docker_restart.py
