+++ calico/adapter/netns.py
@@ -13,19 +13,16 @@
 #    License for the specific language governing permissions and limitations
 #    under the License.
 
-from subprocess import call, check_output, check_call, CalledProcessError
+from subprocess import call, check_output, check_call
 import socket
 import logging
 import logging.handlers
 import sys
 import uuid
-import os.path
 
 from netaddr import IPNetwork, IPAddress
-
-from datastore import Endpoint, IF_PREFIX
+from calico.adapter.datastore import Endpoint, IF_PREFIX
 from nsenter import Namespace
-
 
 _log = logging.getLogger(__name__)
