+++ calico/adapter/requirements.txt
