+++ calico/tests/fv/no_powerstrip.sh
