+++ calico/tests/scale/cloud-config-master.yaml
