+++ calico/tests/scale/cloud-config.yaml
