+++ calico/tests/scale/kill.sh
