+++ node_fs/etc/powerstrip/adapters.yml
