+++ node_fs/etc/service/powerstrip-calico/run
