+++ node_fs/etc/service/powerstrip/run
