+++ node_fs/templates/bird.cfg.template
