+++ node_fs/templates/bird6.cfg.template
