+++ node_fs/templates/felix.cfg.template
